Renaming / moving files in contrib.gauth into gauth_datastore & gauth_sql.
