+++ djangae/contrib/gauth_datastore/__init__.py
