+++ djangae/contrib/gauth_datastore/backends.py
