+++ djangae/contrib/gauth_datastore/permissions.py
