+++ djangae/contrib/gauth_sql/models.py
@@ -20,7 +20,6 @@ class GaeUser(GaeAbstractBaseUser, PermissionsMixin):
     """
 
     class Meta:
-        app_label = "djangae"
         swappable = 'AUTH_USER_MODEL'
         verbose_name = _('user')
         verbose_name_plural = _('users')
